Create the base structure of the project and created the basic components
Created the folders for the files to go in depending on their purpose. Created the main components to run the web application. The components are to create the app itself with the website components and having the necessary stuff to observe and routing processes. In addition to the user model for managing any user processes. The pages created are navigation bar with the login and log out functionality and the home page.