Fixed a bug in error reporting
Fixed a bug in error reporting where only the latest error message would show up only, in case of multiple errors. The reason behind the problem is the breaking changes in React 18. The solution was to downgrade to React 17.
Refactored the code to use the useEffect hook for the error reporting.